+++ src/api/mod.rs
@@ -1,8 +1,8 @@
 pub mod android;
 pub mod caca;
-pub mod dlopen;
 pub mod egl;
 pub mod glx;
 pub mod ios;
 pub mod osmesa;
 pub mod wgl;
+pub mod dlloader;