+++ src/platform/linux/wayland.rs
@@ -1,8 +1,6 @@
-use api::dlopen;
 use api::egl::{self, ffi, Context as EglContext};
-use std::ffi::CString;
 use std::os::raw;
-use std::sync::Arc;
+use std::sync::{Arc, Mutex};
 use wayland_client::egl as wegl;
 use winit;
 use winit::os::unix::WindowExt;
@@ -10,6 +8,11 @@ use {
     ContextError, CreationError, GlAttributes, PixelFormat,
     PixelFormatRequirements,
 };
+use api::dlloader::GlxOrEgl;
+
+lazy_static! {
+    static ref GLX_OR_EGL: Mutex<GlxOrEgl> = Mutex::new(GlxOrEgl::new(false));
+}
 
 pub struct Context {
     egl_surface: Arc<wegl::WlEglSurface>,
@@ -53,34 +56,16 @@ impl Context {
             )
         };
         let context = {
-            let mut libegl = unsafe {
-                dlopen::dlopen(
-                    b"libEGL.so.1\0".as_ptr() as *const _,
-                    dlopen::RTLD_NOW,
-                )
-            };
-            if libegl.is_null() {
-                libegl = unsafe {
-                    dlopen::dlopen(
-                        b"libEGL.so\0".as_ptr() as *const _,
-                        dlopen::RTLD_NOW,
-                    )
-                };
-            }
-            if libegl.is_null() {
-                return Err(CreationError::NotSupported(
-                    "could not find libEGL",
-                ));
-            }
-            let egl = ::api::egl::ffi::egl::Egl::load_with(|sym| {
-                let sym = CString::new(sym).unwrap();
-                unsafe { dlopen::dlsym(libegl, sym.as_ptr()) }
-            });
+            let glx_or_egl = GLX_OR_EGL.lock().unwrap();
+            let egl = glx_or_egl.egl.as_ref().ok_or(
+                CreationError::NotSupported("could not find libEGL")
+            )?;
+
             let gl_attr = gl_attr.clone().map_sharing(|c| &c.context);
             let native_display = egl::NativeDisplay::Wayland(Some(
                 window.get_wayland_display().unwrap() as *const _,
             ));
-            EglContext::new(egl, pf_reqs, &gl_attr, native_display)
+            EglContext::new(egl.clone(), pf_reqs, &gl_attr, native_display)
                 .and_then(|p| p.finish(egl_surface.ptr() as *const _))?
         };
         let context = Context {