+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- We no longer load `libegl.so` and `libgl.so` multiple times.
 - Fixes `Context::is_current` incorrectly returning `false`.
 - **Breaking:** Renamed `GlContext{,Ext}` to `ContextTrait{,Ext}`.
 - Implemented context sharing support for Windows and Linux.